Split Booking Service by creating new CAS3 Booking Service
