+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/BookingService.kt
@@ -1,8 +1,6 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.service
 
-import io.sentry.Sentry
 import jakarta.transaction.Transactional
-import org.slf4j.LoggerFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.BookingStatus
@@ -12,8 +10,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApprovedPremi
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApprovedPremisesEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ArrivalEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ArrivalRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.BedEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.BedRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.BookingEntity
@@ -28,19 +24,11 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ConfirmationE
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ConfirmationRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DateChangeEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DateChangeRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DepartureEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DepartureReasonRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DepartureRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ExtensionEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ExtensionRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.LostBedsRepository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.MoveOnCategoryRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.PlacementRequestRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.PremisesEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.PremisesRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TemporaryAccommodationPremisesEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TurnaroundEntity
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.TurnaroundRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserPermission
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserQualification
@@ -60,45 +48,33 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Withdrawabl
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawalContext
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawalTriggeredBySeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawalTriggeredByUser
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.extractEntityFromCasResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.toLocalDateTime
 import java.time.LocalDate
 import java.time.OffsetDateTime
 import java.util.UUID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas3.DomainEventService as Cas3DomainEventService
 
 @Service
 class BookingService(
   private val premisesService: PremisesService,
   private val offenderService: OffenderService,
-  private val cas3DomainEventService: Cas3DomainEventService,
   private val workingDayService: WorkingDayService,
   private val bookingRepository: BookingRepository,
   private val arrivalRepository: ArrivalRepository,
   private val cancellationRepository: CancellationRepository,
   private val confirmationRepository: ConfirmationRepository,
-  private val extensionRepository: ExtensionRepository,
   private val dateChangeRepository: DateChangeRepository,
-  private val departureRepository: DepartureRepository,
-  private val departureReasonRepository: DepartureReasonRepository,
-  private val moveOnCategoryRepository: MoveOnCategoryRepository,
   private val cancellationReasonRepository: CancellationReasonRepository,
   private val bedRepository: BedRepository,
   private val placementRequestRepository: PlacementRequestRepository,
   private val lostBedsRepository: LostBedsRepository,
-  private val turnaroundRepository: TurnaroundRepository,
   private val premisesRepository: PremisesRepository,
-  private val assessmentRepository: AssessmentRepository,
   private val userService: UserService,
   private val userAccessService: UserAccessService,
-  private val assessmentService: AssessmentService,
   private val cas1BookingEmailService: Cas1BookingEmailService,
   private val deliusService: DeliusService,
   private val cas1BookingDomainEventService: Cas1BookingDomainEventService,
   private val cas1ApplicationStatusService: Cas1ApplicationStatusService,
 ) {
-  private val log = LoggerFactory.getLogger(this::class.java)
-
   fun updateBooking(bookingEntity: BookingEntity): BookingEntity = bookingRepository.save(bookingEntity)
 
   @SuppressWarnings("ThrowsCount")
@@ -252,157 +228,6 @@ class BookingService(
     return AuthorisableActionResult.Success(validationResult)
   }
 
-  @Transactional
-  fun createTemporaryAccommodationBooking(
-    user: UserEntity,
-    premises: TemporaryAccommodationPremisesEntity,
-    crn: String,
-    nomsNumber: String?,
-    arrivalDate: LocalDate,
-    departureDate: LocalDate,
-    bedId: UUID?,
-    assessmentId: UUID?,
-    enableTurnarounds: Boolean,
-  ): AuthorisableActionResult<ValidatableActionResult<BookingEntity>> {
-    val validationResult = validated {
-      if (bedId == null) {
-        "$.bedId" hasValidationError "empty"
-        return@validated fieldValidationError
-      }
-
-      val expectedLastUnavailableDate =
-        workingDayService.addWorkingDays(departureDate, premises.turnaroundWorkingDayCount)
-      getBookingWithConflictingDates(arrivalDate, expectedLastUnavailableDate, null, bedId)?.let {
-        return@validated it.id hasConflictError "A Booking already exists for dates from ${it.arrivalDate} to ${it.lastUnavailableDate} which overlaps with the desired dates"
-      }
-
-      getLostBedWithConflictingDates(arrivalDate, expectedLastUnavailableDate, null, bedId)?.let {
-        return@validated it.id hasConflictError "A Lost Bed already exists for dates from ${it.startDate} to ${it.endDate} which overlaps with the desired dates"
-      }
-
-      bedRepository.findArchivedBedByBedIdAndDate(bedId, departureDate)?.let {
-        return@validated it.id hasConflictError "BedSpace is archived from ${it.endDate} which overlaps with the desired dates"
-      }
-
-      val bed = bedRepository.findByIdOrNull(bedId)
-
-      if (bed == null) {
-        "$.bedId" hasValidationError "doesNotExist"
-      } else if (bed.room.premises !is TemporaryAccommodationPremisesEntity) {
-        "$.bedId" hasValidationError "mustBelongToTemporaryAccommodationPremises"
-      }
-
-      if (departureDate.isBefore(arrivalDate)) {
-        "$.departureDate" hasValidationError "beforeBookingArrivalDate"
-      }
-
-      val application = when (assessmentId) {
-        null -> null
-        else -> {
-          val result = assessmentRepository.findByIdOrNull(assessmentId)
-          if (result == null) {
-            "$.assessmentId" hasValidationError "doesNotExist"
-          }
-          result?.application
-        }
-      }
-
-      if (validationErrors.any()) {
-        return@validated fieldValidationError
-      }
-
-      val bookingCreatedAt = OffsetDateTime.now()
-
-      val booking = bookingRepository.save(
-        BookingEntity(
-          id = UUID.randomUUID(),
-          crn = crn,
-          nomsNumber = nomsNumber,
-          arrivalDate = arrivalDate,
-          departureDate = departureDate,
-          keyWorkerStaffCode = null,
-          arrivals = mutableListOf(),
-          departures = mutableListOf(),
-          nonArrival = null,
-          cancellations = mutableListOf(),
-          confirmation = null,
-          extensions = mutableListOf(),
-          dateChanges = mutableListOf(),
-          premises = premises,
-          bed = bed,
-          service = ServiceName.temporaryAccommodation.value,
-          originalArrivalDate = arrivalDate,
-          originalDepartureDate = departureDate,
-          createdAt = bookingCreatedAt,
-          application = application,
-          offlineApplication = null,
-          turnarounds = mutableListOf(),
-          placementRequest = null,
-          status = BookingStatus.provisional,
-        ),
-      )
-
-      val turnaround = turnaroundRepository.save(
-        TurnaroundEntity(
-          id = UUID.randomUUID(),
-          workingDayCount = when (enableTurnarounds) {
-            true -> premises.turnaroundWorkingDayCount
-            else -> 0
-          },
-          createdAt = bookingCreatedAt,
-          booking = booking,
-        ),
-      )
-
-      booking.turnarounds += turnaround
-
-      cas3DomainEventService.saveBookingProvisionallyMadeEvent(booking, user)
-
-      success(booking)
-    }
-
-    return AuthorisableActionResult.Success(validationResult)
-  }
-
-  @Transactional
-  fun createTurnaround(
-    booking: BookingEntity,
-    workingDays: Int,
-  ) = validated {
-    if (workingDays < 0) {
-      "$.workingDays" hasValidationError "isNotAPositiveInteger"
-    }
-
-    val expectedLastUnavailableDate = workingDayService.addWorkingDays(booking.departureDate, workingDays)
-    getBookingWithConflictingDates(
-      booking.arrivalDate,
-      expectedLastUnavailableDate,
-      booking.id,
-      booking.bed!!.id,
-    )?.let {
-      return@validated it.id hasConflictError "A Booking already exists for dates from ${it.arrivalDate} to ${it.lastUnavailableDate} which overlaps with the desired dates"
-    }
-
-    getLostBedWithConflictingDates(booking.arrivalDate, expectedLastUnavailableDate, null, booking.bed!!.id)?.let {
-      return@validated it.id hasConflictError "A Lost Bed already exists for dates from ${it.startDate} to ${it.endDate} which overlaps with the desired dates"
-    }
-
-    if (validationErrors.any()) {
-      return fieldValidationError
-    }
-
-    val turnaround = turnaroundRepository.save(
-      TurnaroundEntity(
-        id = UUID.randomUUID(),
-        workingDayCount = workingDays,
-        createdAt = OffsetDateTime.now(),
-        booking = booking,
-      ),
-    )
-
-    return success(turnaround)
-  }
-
   @SuppressWarnings("UnusedParameter")
   @Transactional
   fun createArrival(
@@ -443,50 +268,6 @@ class BookingService(
     updateBooking(booking)
 
     booking.arrivals += arrivalEntity
-
-    return success(arrivalEntity)
-  }
-
-  @Transactional
-  fun createCas3Arrival(
-    user: UserEntity,
-    booking: BookingEntity,
-    arrivalDate: LocalDate,
-    expectedDepartureDate: LocalDate,
-    notes: String?,
-    keyWorkerStaffCode: String?,
-  ) = validated<ArrivalEntity> {
-    if (booking.premises !is TemporaryAccommodationPremisesEntity) {
-      return generalError("CAS3 Arrivals cannot be set on non-CAS3 premise")
-    }
-
-    if (expectedDepartureDate.isBefore(arrivalDate)) {
-      return "$.expectedDepartureDate" hasSingleValidationError "beforeBookingArrivalDate"
-    }
-    val isFirstArrival = booking.arrivals.isNullOrEmpty()
-    val arrivalEntity = arrivalRepository.save(
-      ArrivalEntity(
-        id = UUID.randomUUID(),
-        arrivalDate = arrivalDate,
-        arrivalDateTime = arrivalDate.toLocalDateTime().toInstant(),
-        expectedDepartureDate = expectedDepartureDate,
-        notes = notes,
-        booking = booking,
-        createdAt = OffsetDateTime.now(),
-      ),
-    )
-
-    booking.arrivalDate = arrivalDate
-    booking.departureDate = expectedDepartureDate
-    booking.status = BookingStatus.arrived
-    updateBooking(booking)
-
-    booking.arrivals += arrivalEntity
-
-    when (isFirstArrival) {
-      true -> cas3DomainEventService.savePersonArrivedEvent(booking, user)
-      else -> cas3DomainEventService.savePersonArrivedUpdatedEvent(booking, user)
-    }
 
     return success(arrivalEntity)
   }
@@ -607,57 +388,10 @@ class BookingService(
   }
 
   @Transactional
-  fun createCas3Cancellation(
-    booking: BookingEntity,
-    cancelledAt: LocalDate,
-    reasonId: UUID,
-    notes: String?,
-    user: UserEntity,
-  ) = validated<CancellationEntity> {
-    val reason = cancellationReasonRepository.findByIdOrNull(reasonId)
-    if (reason == null) {
-      "$.reason" hasValidationError "doesNotExist"
-    } else if (!reason.serviceScopeMatches(booking.service)) {
-      "$.reason" hasValidationError "incorrectCancellationReasonServiceScope"
-    }
-
-    if (validationErrors.any()) {
-      return fieldValidationError
-    }
-    val isFirstCancellations = booking.cancellations.isNullOrEmpty()
-    val cancellationEntity = cancellationRepository.save(
-      CancellationEntity(
-        id = UUID.randomUUID(),
-        date = cancelledAt,
-        reason = reason!!,
-        notes = notes,
-        booking = booking,
-        createdAt = OffsetDateTime.now(),
-        otherReason = null,
-      ),
-    )
-    booking.status = BookingStatus.cancelled
-    updateBooking(booking)
-    booking.cancellations += cancellationEntity
-
-    when (isFirstCancellations) {
-      true -> cas3DomainEventService.saveBookingCancelledEvent(booking, user)
-      else -> cas3DomainEventService.saveBookingCancelledUpdatedEvent(booking, user)
-    }
-
-    booking.application?.getLatestAssessment()?.let { applicationEntity ->
-      moveTransitionalAccommodationAssessmentToReadyToPlace(user, applicationEntity, booking.id)
-    }
-
-    return success(cancellationEntity)
-  }
-
-  @Transactional
   fun createConfirmation(
     booking: BookingEntity,
     dateTime: OffsetDateTime,
     notes: String?,
-    user: UserEntity,
   ) = validated<ConfirmationEntity> {
     if (booking.confirmation != null) {
       return generalError("This Booking already has a Confirmation set")
@@ -676,133 +410,7 @@ class BookingService(
     updateBooking(booking)
     booking.confirmation = confirmationEntity
 
-    if (booking.premises is TemporaryAccommodationPremisesEntity) {
-      cas3DomainEventService.saveBookingConfirmedEvent(booking, user)
-      findAndCloseCAS3Assessment(booking, user)
-    }
-
     return success(confirmationEntity)
-  }
-
-  fun createDeparture(
-    user: UserEntity,
-    booking: BookingEntity,
-    dateTime: OffsetDateTime,
-    reasonId: UUID,
-    moveOnCategoryId: UUID,
-    notes: String?,
-  ) = when (booking.premises) {
-    is TemporaryAccommodationPremisesEntity ->
-      createCas3Departure(booking, dateTime, reasonId, moveOnCategoryId, notes, user)
-
-    else ->
-      throw RuntimeException("Unknown premises type ${booking.premises::class.qualifiedName}")
-  }
-
-  private fun createCas3Departure(
-    booking: BookingEntity,
-    dateTime: OffsetDateTime,
-    reasonId: UUID,
-    moveOnCategoryId: UUID,
-    notes: String?,
-    user: UserEntity,
-  ) = validated<DepartureEntity> {
-    if (booking.premises !is TemporaryAccommodationPremisesEntity) {
-      throw RuntimeException("Only CAS3 bookings are supported")
-    }
-
-    if (booking.arrivalDate.toLocalDateTime().isAfter(dateTime)) {
-      "$.dateTime" hasValidationError "beforeBookingArrivalDate"
-    }
-
-    val reason = departureReasonRepository.findByIdOrNull(reasonId)
-    if (reason == null) {
-      "$.reasonId" hasValidationError "doesNotExist"
-    } else if (!reason.serviceScopeMatches(booking.service)) {
-      "$.reasonId" hasValidationError "incorrectDepartureReasonServiceScope"
-    }
-
-    val moveOnCategory = moveOnCategoryRepository.findByIdOrNull(moveOnCategoryId)
-    if (moveOnCategory == null) {
-      "$.moveOnCategoryId" hasValidationError "doesNotExist"
-    } else if (!moveOnCategory.serviceScopeMatches(booking.service)) {
-      "$.moveOnCategoryId" hasValidationError "incorrectMoveOnCategoryServiceScope"
-    }
-
-    if (validationErrors.any()) {
-      return fieldValidationError
-    }
-
-    val isFirstDeparture = booking.departures.isNullOrEmpty()
-
-    val departureEntity = departureRepository.save(
-      DepartureEntity(
-        id = UUID.randomUUID(),
-        dateTime = dateTime,
-        reason = reason!!,
-        moveOnCategory = moveOnCategory!!,
-        destinationProvider = null,
-        notes = notes,
-        booking = booking,
-        createdAt = OffsetDateTime.now(),
-      ),
-    )
-    booking.status = BookingStatus.departed
-    booking.departureDate = dateTime.toLocalDate()
-    updateBooking(booking)
-    booking.departures += departureEntity
-
-    when (isFirstDeparture) {
-      true -> cas3DomainEventService.savePersonDepartedEvent(booking, user)
-      else -> cas3DomainEventService.savePersonDepartureUpdatedEvent(booking, user)
-    }
-
-    return success(departureEntity)
-  }
-
-  @Transactional
-  fun createExtension(
-    booking: BookingEntity,
-    newDepartureDate: LocalDate,
-    notes: String?,
-  ) = validated {
-    if (booking.service != ServiceName.temporaryAccommodation.value) {
-      return generalError("Extensions are only supported by CAS3")
-    }
-
-    val expectedLastUnavailableDate =
-      workingDayService.addWorkingDays(newDepartureDate, booking.turnaround?.workingDayCount ?: 0)
-
-    val bedId = booking.bed?.id
-      ?: throw InternalServerErrorProblem("No bed ID present on Booking: ${booking.id}")
-
-    getBookingWithConflictingDates(booking.arrivalDate, expectedLastUnavailableDate, booking.id, bedId)?.let {
-      return@validated it.id hasConflictError "A Booking already exists for dates from ${it.arrivalDate} to ${it.lastUnavailableDate} which overlaps with the desired dates"
-    }
-
-    getLostBedWithConflictingDates(booking.arrivalDate, expectedLastUnavailableDate, null, bedId)?.let {
-      return@validated it.id hasConflictError "A Lost Bed already exists for dates from ${it.startDate} to ${it.endDate} which overlaps with the desired dates"
-    }
-
-    if (booking.arrivalDate.isAfter(newDepartureDate)) {
-      return "$.newDepartureDate" hasSingleValidationError "beforeBookingArrivalDate"
-    }
-
-    val extensionEntity = ExtensionEntity(
-      id = UUID.randomUUID(),
-      previousDepartureDate = booking.departureDate,
-      newDepartureDate = newDepartureDate,
-      notes = notes,
-      booking = booking,
-      createdAt = OffsetDateTime.now(),
-    )
-
-    val extension = extensionRepository.save(extensionEntity)
-    booking.departureDate = extensionEntity.newDepartureDate
-    booking.extensions.add(extension)
-    updateBooking(booking)
-
-    return success(extensionEntity)
   }
 
   private fun shouldCreateDomainEventForBooking(booking: BookingEntity, user: UserEntity?) =
@@ -919,59 +527,6 @@ class BookingService(
 
   val BookingEntity.lastUnavailableDate: LocalDate
     get() = workingDayService.addWorkingDays(this.departureDate, this.turnaround?.workingDayCount ?: 0)
-
-  private fun findAndCloseCAS3Assessment(booking: BookingEntity, user: UserEntity) {
-    booking.application?.let {
-      val assessment =
-        assessmentRepository.findByApplicationIdAndReallocatedAtNull(booking.application!!.id)
-      if (assessment != null) {
-        closeTransitionalAccommodationAssessment(assessment.id, user, booking)
-      }
-    }
-  }
-
-  @SuppressWarnings("TooGenericExceptionCaught")
-  private fun moveTransitionalAccommodationAssessmentToReadyToPlace(
-    user: UserEntity,
-    applicationEntity: AssessmentEntity,
-    bookingId: UUID,
-  ) {
-    try {
-      extractEntityFromCasResult(
-        assessmentService.acceptAssessment(
-          user,
-          applicationEntity.id,
-          applicationEntity.document,
-          null,
-          null,
-          null,
-          "Automatically moved to ready-to-place after booking is cancelled",
-        ),
-      )
-    } catch (exception: Exception) {
-      log.error("Unable to move CAS3 assessment ${applicationEntity.id} to ready-to-place queue for $bookingId ")
-      Sentry.captureException(
-        RuntimeException(
-          "Unable to move CAS3 assessment ${applicationEntity.id} for ready-to-place queue for $bookingId ",
-          exception,
-        ),
-      )
-    }
-  }
-
-  @SuppressWarnings("TooGenericExceptionCaught")
-  private fun closeTransitionalAccommodationAssessment(
-    assessmentId: UUID,
-    user: UserEntity,
-    booking: BookingEntity,
-  ) {
-    try {
-      extractEntityFromCasResult(assessmentService.closeAssessment(user, assessmentId))
-    } catch (exception: Exception) {
-      log.error("Unable to close CAS3 assessment $assessmentId for booking ${booking.id} ", exception)
-      Sentry.captureException(RuntimeException("Unable to close CAS3 assessment $assessmentId for booking ${booking.id} ", exception))
-    }
-  }
 }
 
 sealed interface GetBookingForPremisesResult {
